fix (again) pidfile removing: now the cleanup only happens in the main (initial) process, not in the forked children
Ignore-this: cde76fcaeb30bc50adadfd2b56981373

darcs-hash:20120107092956-161e5-434caf01740dfd2135f13c66205e7d6404b67e4e.gz

+++ home/bin/templates/acis.fcgi.pl
@@ -13,18 +13,20 @@ if (open PID, ">$pidfile") {
   undef $pidfile;
 }
 
-# clean-up pid file after myself:
+# clean-up pid file after myself.  $::acis_forked_child is set in
+# ACIS::Web::Background at forking. Only the main (parent) process
+# should do the cleanup.
 END {  
-  if ($pidfile) { 
-    unlink $pidfile; 
-    undef $pidfile; 
+  if (not $::acis_forked_child and $pidfile) { 
     if ($acis) {
       $acis->log( "END. removing $pidfile" );
     }
+    unlink $pidfile; 
+    undef $pidfile; 
   } 
 }
 local $SIG{TERM} = sub {
-  if ($pidfile) { 
+  if (not $::acis_forked_child and $pidfile) { 
     if ($acis) {
       $acis->log( "TERM signal. removing $pidfile" );
     }
@@ -32,6 +34,8 @@ local $SIG{TERM} = sub {
     undef $pidfile; 
   }  
 };
+
+
 
 ## not sure why we need this
 umask 0000;
